updated role, user classes
