+++ application/classes/models/role.py
@@ -8,6 +8,25 @@ class Role:
     '''
         A parent class for the system roles.
     '''
+
+    def get_roles(self):
+       
+        return self.db_manager.do_select('SELECT * FROM roles')
+       
+
+    def get_role(self, id):
+      
+        if id:
+            return self.db_manager.do_select('SELECT * FROM roles WHERE id = ?', (id,) )
+        else:
+            return False
+        
+    def get_role_id(self, name):
+        result = self.db_manager.do_select('SELECT id FROM roles WHERE name = ?', (name,) )
+        if len(result) >= 1:        
+            return result[0]['name'] # TODO
+        else:
+            return 0
     
     # Creates a record and returns the inserted id
     def add_role(self, name):
@@ -26,7 +45,16 @@ class Role:
         query = "DELETE FROM roles WHERE id = ?"
         where = (name,)
         # call do_delete method from DBManager
-        return self.db_manager.do_delete(query, where, False)   
+        return self.db_manager.do_delete(query, where, False) 
+      
+    
+    def update_role(self, id, name):
+        # perform database query to update permission attributes.
+        query = "UPDATE roles SET name='" + name + "', updated_at= " + self._updated_at + " WHERE role_id=?"
+        values = (id,)
+
+        # call do_update method from DBmanager.
+        self.db_manager.do_update(query, values)
     
 
     def connect_db_manager(self, db_manager):
